+++ R/ns.R
@@ -19,7 +19,7 @@
 #' @details `step_ns` can create new features from a single variable
 #'  that enable fitting routines to model this variable in a
 #'  nonlinear manner. The extent of the possible nonlinearity is
-#'  determined by the `df` or `knot` arguments of
+#'  determined by the `df` or `knots` arguments of
 #'  [splines::ns()]. The original variables are removed
 #'  from the data and new columns are added. The naming convention
 #'  for the new variables is `varname_ns_1` and so on.
@@ -117,12 +117,17 @@ ns_statistics <- function(x, args) {
     ok <- !is.na(x) & x >= boundary[1L] & x <= boundary[2L]
     knots <- unname(quantile(x[ok], seq_len(num_knots) / (num_knots + 1L)))
   } else {
-    knots <- numeric()
+    if (is.null(args$knots)) {
+      knots <- numeric()
+    } else {
+      knots <- args$knots
+    }
   }
 
   # Only construct the data necessary for splines_predict
   out <- matrix(NA, ncol = degree + length(knots) + intercept, nrow = 1L)
   class(out) <- c("ns", "basis", "matrix")
+  attr(out, "degree") <- 3L
   attr(out, "knots") <- knots
   attr(out, "Boundary.knots") <- boundary
   attr(out, "intercept") <- intercept